Add files via upload
Removed loop function from main starting video player, it would eventually crash but not fully quit, leaving the system in a half-running state without the knowledge of the main thread handler. This means it is no longer seamless looping (the system kick starts the player each time it finishes), but will be reliable now.